Add hybrid gpu checks with notice; Refactor steamvr_launcher to move out platform-specific calls to separate modules
